+++ storetaskid.php
@@ -2,5 +2,4 @@
 	if (isset($_POST['taskid'])) {
 		echo json_encode(array("abc" => $_POST['taskid']));
 	}
-
 ?>
